+++ cg/cli/generate/report/utils.py
